Update InventoryItem fields/record_refs for 2021.2 (#506)
A couple were missing for my use case, so I updated them based on 2021.2

`member_list` was removed as a field as it doesn't exist on
InventoryItem in the NetSuite schema. It was added back in a108ae4d to
appease searching, at a time when KitItem inherited from InventoryItem,
which is no longer the case. If I'm understanding the ItemSearchRowBasic
schema correctly, it's not returned as a field there:
https://system.netsuite.com/help/helpcenter/en_US/srbrowser/Browser2021_2/schema/search/itemsearchrowbasic.html?mode=package