Remove monkey patched `lower_camelcase` method on String (#533)
Our company enforces that we do not depend on gems that monkey patch core Ruby classes. To enable us to depend on this gem, this change moves the `lower_camelcase` to be a static method on a new utility class `NetSuite::Utilities::Strings`.

There are many places where `lower_camelcase` is used to generate a `record_type` so this change consolidates the logic in `NetSuite::Support::Records` and uses it where a record type is expected.