Avoid Savon version `2.13.0` to prevent generating invalid envelopes (#563)
* Avoid Savon version `2.13.0` to prevent generating invalid envelopes

This version of Savon mistakenly prefixed the default namespace of
`xmlns`, resulting in a namespace of `xmlns:xmlns`, which caused
NetSuite to error on requests. v2.13.1 fixed this issue to skip
prefixing the default namespace.

This is a continuation of #558, which avoided v2.13.0 and anything
greater.

Bug: https://github.com/savonrb/savon/pull/943#issuecomment-1213163418
Fix: https://github.com/savonrb/savon/pull/977

* Use RC of mail gem in CI for Ruby 3.1 support

Savon v2.13 adds a dependency on mail, which has an implicit dependency
on net-smtp. Ruby 3.1 removed net-smtp as a default gem. mail v2.8.0.rc1
is the first to make that dependency explicit to support Ruby 3.1.

This at least gets CI passing, but may not help end-users on Ruby 3.1.
If they previously had a dependency on mail (ie. through Rails), there's
probably already an explicit dependency in their app on net-smtp to get
it working on Ruby 3.1. If they have no dependency on mail previously,
this will likely error as they go to require this gem, which requires
savon, which requires mail, which requires net-smtp, but Ruby can't find
it. It doesn't seem right to add net-smtp (or an RC of mail) to this
gem's gemspec dependencies, given we don't directly use either library.
Ultimately it's about the latest stable release of mail being
incompatible with Ruby 3.1.