Optionally pass in `proxy` parameter to Savon (#548)
* Optionally pass in `proxy` parameter to Savon

Savon will call `to_s` on any `proxy` value passed in so it will treat `nil` like an empty string instead of not using it. This causes requests to fail with an invalid proxy. This change will not pass in a value for `proxy` if the attribute is `nil`.

This change also updates history with addition of the proxy attribute.

* Updated PR with cleaner way to add the proxy and added testing

---------

Co-authored-by: Michael Bianco <iloveitaly@gmail.com>